// src/models/Property.ts
import { AcType } from "./AcType";
import { Area } from "./Area";
import { City } from "./City";
import { PaymentMethod } from "./PaymentMethod";
import { PropertyCategory } from "./PropertyCategory";
import { PropertyType } from "./PropertyType";
import { User } from "./User";

export type Property = {
  id: number;
  propertyTypeId: number;
  propertyCategoryId: number;
  userId: number;
  cityId: number;
  areaId: number;
  acTypeId: number;
  paymentMethodId: number;
  featured: boolean;
  title: string;
  description: string;
  imageNames: string;
  externalImage: string;
  bedRoomCount: number;
  bathRoomCount: number;
  kitchenRoomCount: number;
  parkingCount: number;
  additionalFeatures: string;
  price: number;
  currency: string;
  address: string;
  latitude: number;
  longitude: number;
  createdDate: Date;
  city: City;
  area: Area;
  acTye: AcType;
  paymentMethod: PaymentMethod;
  propertyCategory: PropertyCategory;
  propertyType: PropertyType;
  user: User;
  size: number;
  maximum: number;
  isLiked: boolean;
  visible: boolean;
};